Rather then looking for a ~ at the end of a escape sequence, which misses [A, Z] (and [a, z]), Pl_LE_Get_Char not looks continues while the read character is ; or in [0, 9]. ; i treated as digit 10 and the read characters are encoding in base 11, rather than in base 10.
This together with that the last character is included in 8 bits, and
the words 31:th bit is set, allows all ANSI CSI characters to be encoded,
assuming thare are non that contains anything else than ; and [0, 9], but
not of those are recognised, so that does not matters.

Additional F1, F2, F3, F4, Home and End has alternative encodings,
starting with ESC O followed by a letters, P, Q, R, S, H, F, respectively,
these are these are recognised and stored as encoded CSI sequences by
Pl_LE_Get_Char.

The combined with fixed (as well as adapted for this encoding) key combination
constants allows F1 to F12, and CTRL-arrow to be recognised and the program
not to freeze until ~ is typed. This allows allowed to F1 to F12, arrows,
Home, End, Page Up, Page Down, Insert and Delete to be recogined with any
modifier combination supported by the used terminal. However F1 to F4, Home
and End are not supported by many terminals to be combined with modifiers.